Moved SCNGeometry extensions to a separate extension file
